requirements.txt
ctransformers>=0.2.27
llama-cpp-python>=0.2.80
huggingface_hub>=0.24.6
mysql-connector-python>=8.4.0
